+++ examples/typescript/webpack.config.babel.js
@@ -11,48 +11,52 @@ module.exports = {
       // Required to support async/await
       '@babel/polyfill',
     ],
-    main: [
-      './src/index',
-    ],
+    main: ['./src/index'],
   },
   output: {
     path: path.join(__dirname, 'dist'),
     filename: '[name].js',
   },
   devtool: false,
-  resolve: { extensions: [".ts", ".tsx", ".js", ".jsx"] },
+  resolve: {
+    extensions: ['.ts', '.tsx', '.js', '.jsx'],
+    alias: {
+      'react-hot-loader': path.resolve(path.join(__dirname, './../../')),
+      react: path.resolve(path.join(__dirname, './node_modules/react')),
+    },
+  },
   module: {
     rules: [
       {
         test: /\.(j|t)sx?$/,
         exclude: /node_modules/,
         use: {
-          loader: "babel-loader",
+          loader: 'babel-loader',
           options: {
             cacheDirectory: true,
             babelrc: false,
             presets: [
               [
-                "@babel/preset-env",
-                { targets: { browsers: "last 2 versions" } } // or whatever your project requires
+                '@babel/preset-env',
+                { targets: { browsers: 'last 2 versions' } }, // or whatever your project requires
               ],
-              "@babel/preset-typescript",
-              "@babel/preset-react"
+              '@babel/preset-typescript',
+              '@babel/preset-react',
             ],
             plugins: [
               // plugin-proposal-decorators is only needed if you're using experimental decorators in TypeScript
-              ["@babel/plugin-proposal-decorators", { legacy: true }],
-              ["@babel/plugin-proposal-class-properties", { loose: true }],
-              "react-hot-loader/babel",
-            ]
-          }
-        }
-      }
-    ]
+              ['@babel/plugin-proposal-decorators', { legacy: true }],
+              ['@babel/plugin-proposal-class-properties', { loose: true }],
+              'react-hot-loader/babel',
+            ],
+          },
+        },
+      },
+    ],
   },
   plugins: [
     new ForkTsCheckerWebpackPlugin(),
     new webpack.NamedModulesPlugin(),
     new HtmlWebpackPlugin(),
-  ]
+  ],
 }